+++ package.json
@@ -2,7 +2,7 @@
   "name": "telegra.ph",
   "version": "1.0.1",
   "description": "Tiny API helper for Telegra.ph",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "eslint . && ava test"
   },
@@ -25,8 +25,7 @@
   },
   "homepage": "https://github.com/telegraf/telegra.ph#readme",
   "files": [
-    "index.js",
-    "typings/*.ts"
+    "dist"
   ],
   "devDependencies": {
     "ava": "^1.2.0",
@@ -39,6 +38,7 @@
     "eslint-plugin-standard": "^4.0.0"
   },
   "dependencies": {
+    "@types/isomorphic-fetch": "^0.0.39",
     "isomorphic-fetch": "^2.2.1"
   }
 }
