+++ typings/account.ts
@@ -1,0 +1,8 @@
+export interface Account {
+  short_name: string
+  author_name: string
+  author_url?: string
+  access_token?: string
+  auth_url?: string
+  page_count?: number
+}
