+++ typings/node.ts
@@ -1,0 +1,3 @@
+import { NodeElement } from "./nodeElement"
+
+export type Node = string | NodeElement
