+++ typings/nodeElement.ts
@@ -1,0 +1,32 @@
+export interface NodeElement {
+  tag:
+    | "a"
+    | "aside"
+    | "b"
+    | "blockquote"
+    | "br"
+    | "code"
+    | "em"
+    | "figcaption"
+    | "figure"
+    | "h3"
+    | "h4"
+    | "hr"
+    | "i"
+    | "iframe"
+    | "img"
+    | "li"
+    | "ol"
+    | "p"
+    | "pre"
+    | "s"
+    | "strong"
+    | "u"
+    | "ul"
+    | "video"
+  attrs?: {
+    href?: string
+    src?: string
+  }
+  children?: Array<Node> | string
+}
