+++ typings/pageList.ts
@@ -1,0 +1,5 @@
+import type { Page } from "./page"
+export interface PageList {
+  total_count: number
+  pages: Array<Page>
+}
