+++ typings/pageViews.ts
@@ -1,0 +1,3 @@
+export interface PageViews {
+  views: number
+}
